Added keep-alive support and modified HttpTransport API for HTTP/2
Huge commit. The Transport API now gives chunks of data to the client
rather than the client reading from the transport. This allows for
keep-alive support, where a single HttpClient always serves a single
request. It also accounts for the forseeable HTTP/2 support, which will
require multiplexing support in the Transport somehow.
Another thing is that HttpClient now supports Streaming (Which is the
old behaviour, thus, no keep-alive support), ChunkedStreaming (Which
uses chunked transfer, default for keep-alive connections) and Buffered
which buffers all writes to the HttpClient internally and then sends
it all at once.