Broke filter behaviour to keep compat with the world (Content-Encoding)
So, when I implemented compression I looked up on how to correctly tell
the client that the body is indeed compressed using gzip or deflate.
Turned out that the correct way is the Transfer-Encoding header, which
was then used. Earlier today though, I heard that some less compliant
implementations of HTTP/1.1 (QNetworkAccessManager) only looks for
Content-Encoding header instead fo the Transfer-Encoding one. After
a discussion it was decided to keep Qts broken behaviour.

I'm thus doing the wrong thing and break HttpClient to use
Content-Encoding to tell the client-side about compression.